+++ src/cpArbiter.h
@@ -40,11 +40,11 @@ typedef struct cpContact{
 	cpFloat bias;
 	
 	// Hash value used to (mostly) uniquely identify a contact.
-	unsigned int hash;
+	size_t hash;
 } cpContact;
 
 // Contacts are always allocated in groups.
-cpContact* cpContactInit(cpContact *con, cpVect p, cpVect n, cpFloat dist, unsigned int hash);
+cpContact* cpContactInit(cpContact *con, cpVect p, cpVect n, cpFloat dist, size_t hash);
 
 // Sum the contact impulses. (Can be used after cpSpaceStep() returns)
 cpVect cpContactsSumImpulses(cpContact *contacts, int numContacts);
